Aggiunto messaggio errore nel caso di admin che tenta di eliminare il suo account

+++ PHP/Elimina utente.php
@@ -1,6 +1,10 @@
 <?php
 require_once('config.php');
 register('user');
-query("DELETE FROM utente WHERE CodiceUtente=$user");
-header("Location:Lista utenti.php");
+if($_SESSION['user_code'] == $user) {
+	header("Location:Errore eliminazione utente.php");
+} else {
+	query("DELETE FROM utente WHERE CodiceUtente=$user");
+	header("Location:Lista utenti.php");
+}
 ?>
